Add base class for options.
