+++ gcovr/writer/html/__init__.py
@@ -22,7 +22,15 @@ import os
 import hashlib
 import io
 from argparse import ArgumentTypeError
-from typing import Any, Callable, Dict, Optional, Union
+from typing import Any, Callable, Dict, List, Optional, Union
+
+from ...options import (
+    GcovrConfigOption,
+    OutputOrDefault,
+    check_input_file,
+    check_percentage,
+)
+from ...writer.base import writer_base
 
 from ...version import __version__
 from ...utils import (
@@ -48,6 +56,224 @@ from ...coverage import (
 )
 
 LOGGER = logging.getLogger("gcovr")
+
+
+class writer(writer_base):
+    def get_options() -> List[GcovrConfigOption]:
+        return [
+            GcovrConfigOption(
+                "html",
+                ["--html"],
+                group="output_options",
+                metavar="OUTPUT",
+                help="Generate a HTML report. OUTPUT is optional and defaults to --output.",
+                nargs="?",
+                type=OutputOrDefault,
+                default=None,
+                const=OutputOrDefault(None),
+            ),
+            GcovrConfigOption(
+                "html_details",
+                ["--html-details"],
+                group="output_options",
+                metavar="OUTPUT",
+                help=(
+                    "Add annotated source code reports to the HTML report. "
+                    "Implies --html, can not be used together with --html-nested. "
+                    "OUTPUT is optional and defaults to --output."
+                ),
+                nargs="?",
+                type=OutputOrDefault,
+                default=None,
+                const=OutputOrDefault(None),
+            ),
+            GcovrConfigOption(
+                "html_nested",
+                ["--html-nested"],
+                group="output_options",
+                metavar="OUTPUT",
+                help=(
+                    "Add annotated source code reports to the HTML report. "
+                    "A page is created for each directory that summarize subdirectories "
+                    "with aggregated statistics. "
+                    "Implies --html, can not be used together with --html-details. "
+                    "OUTPUT is optional and defaults to --output."
+                ),
+                nargs="?",
+                type=OutputOrDefault,
+                default=None,
+                const=OutputOrDefault(None),
+            ),
+            GcovrConfigOption(
+                "html_syntax_highlighting",
+                ["--html-syntax-highlighting", "--html-details-syntax-highlighting"],
+                group="output_options",
+                help="Use syntax highlighting in HTML source page. Enabled by default.",
+                action="store_const",
+                default=True,
+                const=True,
+                const_negate=False,  # autogenerates --no-NAME with action const=False
+            ),
+            GcovrConfigOption(
+                "html_theme",
+                ["--html-theme"],
+                group="output_options",
+                type=str,
+                choices=CssRenderer.get_themes(),
+                metavar="THEME",
+                help=(
+                    "Override the default color theme for the HTML report. "
+                    "Default is {default!s}."
+                ),
+                default=CssRenderer.get_default_theme(),
+            ),
+            GcovrConfigOption(
+                "html_css",
+                ["--html-css"],
+                group="output_options",
+                type=check_input_file,
+                metavar="CSS",
+                help="Override the default style sheet for the HTML report.",
+                default=None,
+            ),
+            GcovrConfigOption(
+                "html_title",
+                ["--html-title"],
+                group="output_options",
+                metavar="TITLE",
+                help="Use TITLE as title for the HTML report. Default is '{default!s}'.",
+                default="GCC Code Coverage Report",
+            ),
+            GcovrConfigOption(
+                "html_medium_threshold",
+                ["--html-medium-threshold"],
+                group="output_options",
+                type=check_percentage,
+                metavar="MEDIUM",
+                help=(
+                    "If the coverage is below MEDIUM, the value is marked "
+                    "as low coverage in the HTML report. "
+                    "MEDIUM has to be lower than or equal to value of --html-high-threshold "
+                    "and greater than 0. "
+                    "If MEDIUM is equal to value of --html-high-threshold the report has "
+                    "only high and low coverage. Default is {default!s}."
+                ),
+                default=75.0,
+            ),
+            GcovrConfigOption(
+                "html_high_threshold",
+                ["--html-high-threshold"],
+                group="output_options",
+                type=check_percentage,
+                metavar="HIGH",
+                help=(
+                    "If the coverage is below HIGH, the value is marked "
+                    "as medium coverage in the HTML report. "
+                    "HIGH has to be greater than or equal to value of --html-medium-threshold. "
+                    "If HIGH is equal to value of --html-medium-threshold the report has "
+                    "only high and low coverage. Default is {default!s}."
+                ),
+                default=90.0,
+            ),
+            GcovrConfigOption(
+                "html_medium_threshold_branch",
+                ["--html-medium-threshold-branch"],
+                group="output_options",
+                metavar="MEDIUM_BRANCH",
+                type=check_percentage,
+                help="If the coverage is below MEDIUM_BRANCH, the value is marked "
+                "as low coverage in the HTML report. "
+                "MEDIUM_BRANCH has to be lower than or equal to value of --html-high-threshold-branch "
+                "and greater than 0. "
+                "If MEDIUM_BRANCH is equal to value of --html-medium-threshold-branch the report has "
+                "only high and low coverage. Default is taken from --html-medium-threshold.",
+                default=None,
+            ),
+            GcovrConfigOption(
+                "html_high_threshold_branch",
+                ["--html-high-threshold-branch"],
+                group="output_options",
+                type=check_percentage,
+                metavar="HIGH_BRANCH",
+                help="If the coverage is below HIGH_BRANCH, the value is marked "
+                "as medium coverage in the HTML report. "
+                "HIGH_BRANCH has to be greater than or equal to value of --html-medium-threshold-branch. "
+                "If HIGH_BRANCH is equal to value of --html-medium-threshold-branch the report has "
+                "only high and low coverage. Default is taken from --html-high-threshold.",
+                default=None,
+            ),
+            GcovrConfigOption(
+                "html_medium_threshold_line",
+                ["--html-medium-threshold-line"],
+                group="output_options",
+                metavar="MEDIUM_LINE",
+                type=check_percentage,
+                help="If the coverage is below MEDIUM_LINE, the value is marked "
+                "as low coverage in the HTML report. "
+                "MEDIUM_LINE has to be lower than or equal to value of --html-high-threshold-line "
+                "and greater than 0. "
+                "If MEDIUM_LINE is equal to value of --html-medium-threshold-line the report has "
+                "only high and low coverage. Default is taken from --html-medium-threshold.",
+                default=None,
+            ),
+            GcovrConfigOption(
+                "html_high_threshold_line",
+                ["--html-high-threshold-line"],
+                group="output_options",
+                type=check_percentage,
+                metavar="HIGH_LINE",
+                help="If the coverage is below HIGH_LINE, the value is marked "
+                "as medium coverage in the HTML report. "
+                "HIGH_LINE has to be greater than or equal to value of --html-medium-threshold-line. "
+                "If HIGH_LINE is equal to value of --html-medium-threshold-line the report has "
+                "only high and low coverage. Default is taken from --html-high-threshold.",
+                default=None,
+            ),
+            GcovrConfigOption(
+                "html_tab_size",
+                ["--html-tab-size"],
+                group="output_options",
+                help="Used spaces for a tab in a source file. Default is {default!s}",
+                type=int,
+                default=4,
+            ),
+            GcovrConfigOption(
+                "relative_anchors",
+                ["--html-absolute-paths"],
+                group="output_options",
+                help=(
+                    "Use absolute paths to link the --html-details reports. "
+                    "Defaults to relative links."
+                ),
+                action="store_false",
+            ),
+            GcovrConfigOption(
+                "html_encoding",
+                ["--html-encoding"],
+                group="output_options",
+                help=(
+                    "Override the declared HTML report encoding. "
+                    "Defaults to {default!s}. "
+                    "See also --source-encoding."
+                ),
+                default="UTF-8",
+            ),
+            GcovrConfigOption(
+                "html_self_contained",
+                ["--html-self-contained"],
+                group="output_options",
+                help=(
+                    "Control whether the HTML report bundles resources like CSS styles. "
+                    "Self-contained reports can be sent via email, "
+                    "but conflict with the Content Security Policy of some web servers. "
+                    "Defaults to self-contained reports unless --html-details is used."
+                ),
+                action="store_const",
+                default=None,
+                const=True,
+                const_negate=False,
+            ),
+        ]
 
 
 class Lazy:
